move detail_levels into converter_gis package
to fix circular import problem and because it kinda belongs there
