+++ osmaxx/conversion/converters/converter_gis/bootstrap/bootstrap.py
@@ -5,9 +5,9 @@ from memoize import mproperty
 
 from osmaxx.conversion.converters.converter_gis.helper.default_postgres import get_default_postgres_wrapper
 from osmaxx.conversion.converters.converter_gis.helper.osm_boundaries_importer import OSMBoundariesImporter
-from osmaxx.conversion.converters import detail_levels
+from osmaxx.conversion.converters.converter_gis import detail_levels
 from osmaxx.conversion.converters.converter_pbf.to_pbf import cut_pbf_along_polyfile
-from osmaxx.conversion.converters.detail_levels import DETAIL_LEVEL_TABLES
+from osmaxx.conversion.converters.converter_gis.detail_levels import DETAIL_LEVEL_TABLES
 from osmaxx.conversion.converters.utils import logged_check_call
 from osmaxx.utils import polyfile_helpers
