+++ osmaxx/conversion/converters/converter_gis/detail_levels.py
