+++ osmaxx/excerptexport/forms/order_options_mixin.py
@@ -3,7 +3,7 @@ from django import forms
 
 from crispy_forms.layout import Fieldset, Div
 
-from osmaxx.conversion.converters.detail_levels import DETAIL_LEVEL_CHOICES
+from osmaxx.conversion.converters.converter_gis.detail_levels import DETAIL_LEVEL_CHOICES
 from osmaxx.conversion_api import formats, coordinate_reference_systems as crs
