+++ tests/conversion/converters/bootstrap_test.py
@@ -1,6 +1,6 @@
 from unittest import mock
 
-from osmaxx.conversion.converters.detail_levels import DETAIL_LEVEL_REDUCED
+from osmaxx.conversion.converters.converter_gis.detail_levels import DETAIL_LEVEL_REDUCED
 from tests.conftest import area_polyfile_string
 
 from osmaxx.conversion.converters.converter_gis.bootstrap import bootstrap
